Fix current source and power source GUI bug, GridDescriptor cleanup

+++ src/main/kotlin/mods/eln/sixnode/CurrentSourceSix.kt
@@ -211,7 +211,7 @@ class CurrentSourceRender(tileEntity: SixNodeEntity, side: Direction, descriptor
     override fun publishUnserialize(stream: DataInputStream) {
         super.publishUnserialize(stream)
         try {
-            current = stream.readFloat().toDouble()
+            current = stream.readDouble()
             needRedrawCable()
         } catch (e: IOException) {
             e.printStackTrace()
